requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
streamlit==1.40.1
openai==1.57.2
google-generativeai==0.8.3
Pillow==11.0.0
pydantic==2.10.3
python-multipart==0.0.19